refactor: Consolidate root directory selection into settings modal
