+++ .changeset/green-garlics-cheer.md
@@ -1,0 +1,5 @@
+---
+"specif-ai": patch
+---
+
+refactor: Consolidate root directory selection into settings modal
