--- ui/src/app/components/select-root-directory/select-root-directory.component.html
@@ -1,35 +1,0 @@
-<div class="modal-wrapper">
-  <div class="modal-header flex items-center justify-between p-4 border-b">
-    <h4 class="text-lg font-medium">Choose Destination Folder</h4>
-    <button
-      type="button"
-      class="text-gray-400 bg-transparent hover:bg-gray-200 hover:text-gray-900 rounded-lg text-sm p-1.5 ml-auto inline-flex items-center focus:outline-none focus:ring-0"
-      data-dismiss="modal"
-      aria-label="Close"
-      (click)="closeModal()"
-    >
-      <ng-icon name="heroXMark"></ng-icon>
-    </button>
-  </div>
-  <div class="modal-body p-6">
-    <div class="flex w-full">
-      <input
-        type="text"
-        [value]="workingDir"
-        placeholder="Choose Destination Folder"
-        disabled
-        class="bg-gray-100 border border-gray-300 text-gray-400 text-sm rounded-l-lg block w-full p-2"
-      />
-      <app-button
-        [buttonContent]="'Browse'"
-        [theme]="'primary'"
-        [rounded]="'none'"
-        [roundedRight]="'lg'"
-        (click)="openFolderSelector()"
-      ></app-button>
-    </div>
-    <p class="mt-4 text-sm text-gray-500">
-      Select a folder to store the requirements generated by {{appName}}
-    </p>
-  </div>
-</div>
