+++ ui/src/app/components/settings/settings.component.html
@@ -1,9 +1,9 @@
-<div class="modal-wrapper w-[32rem] max-w-lg text-sm font-medium">
+<div class="modal-wrapper w-[30rem] max-w-lg text-sm font-medium">
   <div class="modal-header flex items-center justify-between px-4 pt-4">
     <h4 class="font-medium text-base">Settings</h4>
     <button
       type="button"
-      class="text-slate-400 bg-transparent hover:bg-slate-200 hover:text-slate-900 rounded-lg text-sm p-1.5 ml-auto inline-flex items-center focus:outline-none focus:ring-0"
+      class="text-slate-400 bg-transparent hover:bg-slate-200 hover:text-slate-900 rounded-lg p-1 ml-auto inline-flex items-center focus:outline-none focus:ring-0"
       data-dismiss="modal"
       aria-label="Close"
       (click)="closeModal()"
@@ -91,6 +91,30 @@
             ></path>
           </svg>
         </div>
+      </div>
+    </div>
+
+    <!-- Destination Folder -->
+    <div class="space-y-1">
+      <label for="workingDir" class="block text-slate-500">
+        Folder to store the requirements generated by {{ appName }}
+      </label>
+      <div class="flex w-full">
+        <input
+          type="text"
+          [value]="workingDir"
+          placeholder="Choose Destination Folder"
+          disabled
+          id="workingDir"
+          class="bg-gray-100 border border-gray-300 text-gray-400 text-sm rounded-l-lg block w-full p-2"
+        />
+        <app-button
+          [buttonContent]="'Browse'"
+          [theme]="'primary'"
+          [rounded]="'none'"
+          [roundedRight]="'lg'"
+          (click)="openFolderSelector()"
+        ></app-button>
       </div>
     </div>
